Lever working as intended! added Lever.cs, set the values for start&end Orientation as well as Rotation, used Quaternion.Slerp to rotate the lever. These values are also set in the editor and a rigidbody is added to the LeverHandle. Finally, Hand Control script is added to Lever so that hands become hidden when the interactable is grabbed.  Drag events are added in editor for the Drag Interactable scirpt.
added Lever.cs, set the values for start&end Orientation as well as Rotation, used Quaternion.Slerp to rotate the lever. These values are also set in the editor and a rigidbody is added to the LeverHandle. Finally, Hand Control script is added to Lever so that hands become hidden when the interactable is grabbed.

Drag events are added in editor for the Drag Interactable scirpt. Lever working as intended